moving single use functions back from utils to DynodeRunner
